+++ R/data.R
@@ -1,0 +1,12 @@
+#' Estimated Residential Population 2023
+#'
+#' (Estimated) population data at the SA2 level (2021 edition) from the ABS.
+#'
+#' @format `sa2_erp23`
+#' A data frame with 2,472 rows and 2 columns:
+#' \describe{
+#'   \item{sa2_code}{SA2 code (2021 edition)}
+#'   \item{erp}{estimated number of residents}
+#' }
+#' @source <https://www.abs.gov.au/statistics/people/population/regional-population/latest-release#data-downloads>
+"sa2_erp23"